+++ phenotyping/_colony_profile_cell_profiler_integration.py
@@ -12,6 +12,7 @@ console_handler.setFormatter(formatter)
 
 from copy import deepcopy
 import pandas as pd
+import numpy as np
 
 from ..cellprofiler_api import CellProfilerApi
 from ._colony_profile_measure import ColonyProfileMeasure
@@ -25,9 +26,20 @@ class ColonyProfileCellProfilerIntegration(ColonyProfileMeasure):
     Due to how the API works, this class should be the 2nd to last endpoint for the ColonyProfile Class
     """
 
-    def __init__(self, img, sample_name, auto_run=True):
+    def __init__(self, img: np.ndarray, sample_name: str,
+                 auto_run: bool = True,
+                 use_boosted_mask: bool = True,
+                 boost_kernel_size: bool = None,
+                 boost_footprint_radius: bool = 5):
         self._cp_results = None
-        super().__init__(img=img, sample_name=sample_name, auto_run=auto_run)
+        super().__init__(
+                img=img,
+                sample_name=sample_name,
+                auto_run=auto_run,
+                use_boosted_mask=use_boosted_mask,
+                boost_kernel_size=boost_kernel_size,
+                boost_footprint_radius=boost_footprint_radius
+        )
 
     @property
     def _results(self):