+++ phenotyping/_colony_profile_measure_base.py
@@ -2,8 +2,8 @@ import sys
 import logging
 
 formatter = logging.Formatter(
-    fmt=f'[%(asctime)s|%(name)s] %(levelname)s - %(message)s',
-    datefmt='%m/%d/%Y %I:%M:%S'
+        fmt=f'[%(asctime)s|%(name)s] %(levelname)s - %(message)s',
+        datefmt='%m/%d/%Y %I:%M:%S'
 )
 console_handler = logging.StreamHandler()
 log = logging.getLogger(__name__)
@@ -11,14 +11,27 @@ log.addHandler(console_handler)
 console_handler.setFormatter(formatter)
 
 import pandas as pd
+import numpy as np
 
 from ._colony_profile_plot_object import ColonyProfilePlotObject
 
 
 class ColonyProfileMeasureBase(ColonyProfilePlotObject):
-    def __init__(self, img, sample_name, auto_run=True):
+    def __init__(self, img: np.ndarray, sample_name: str,
+                 auto_run: bool = True,
+                 use_boosted_mask: bool = True,
+                 boost_kernel_size: bool = None,
+                 boost_footprint_radius: bool = 5
+                 ):
         self._measurements = {}
-        super().__init__(img=img, sample_name=sample_name, auto_run=auto_run)
+        super().__init__(
+                img=img,
+                sample_name=sample_name,
+                auto_run=auto_run,
+                use_boosted_mask=use_boosted_mask,
+                boost_kernel_size=boost_kernel_size,
+                boost_footprint_radius=boost_footprint_radius
+        )
 
     def run_analysis(self,
                      threshold_method="otsu", use_boosted=True,